// src/lib/proto/RCRSProto.ts
export interface IntListProto {
  values: number[];
}

export interface FloatListProto {
  values: number[];
}

export interface IntMatrixProto {
  values: IntListProto[];
}

export interface ValueProto {
  defined: boolean;
  intValue?: number | undefined;
  boolValue?: boolean | undefined;
  doubleValue?: number | undefined;
  byteList?: Uint8Array | undefined;
  intList?: IntListProto | undefined;
  intMatrix?: IntMatrixProto | undefined;
  edgeList?: EdgeListProto | undefined;
  point2D?: Point2DProto | undefined;
}

export interface PropertyProto {
  urn: number;
  defined: boolean;
  intValue?: number | undefined;
  boolValue?: boolean | undefined;
  doubleValue?: number | undefined;
  byteList?: Uint8Array | undefined;
  intList?: IntListProto | undefined;
  intMatrix?: IntMatrixProto | undefined;
  edgeList?: EdgeListProto | undefined;
  point2D?: Point2DProto | undefined;
}

export interface Point2DProto {
  X: number;
  Y: number;
}

export interface EntityProto {
  urn: number;
  entityID: number;
  properties: PropertyProto[];
}

export interface EntityListProto {
  entities: EntityProto[];
}

export interface ConfigProto {
  data: { [key: string]: string };
}

export interface ConfigProto_DataEntry {
  key: string;
  value: string;
}

export interface EdgeListProto {
  edges: EdgeProto[];
}

export interface EdgeProto {
  startX: number;
  startY: number;
  endX: number;
  endY: number;
  neighbour: number;
}

export interface ChangeSetProto {
  changes: ChangeSetProto_EntityChangeProto[];
  deletes: number[];
}

export interface ChangeSetProto_EntityChangeProto {
  entityID: number;
  urn: number;
  properties: PropertyProto[];
}
// ...
export const IntListProto: MessageFns<IntListProto> = {
  encode(message: IntListProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    writer.uint32(10).fork();
    for (const v of message.values) {
      writer.int32(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IntListProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIntListProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag === 8) {
            message.values.push(reader.int32());

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.values.push(reader.int32());
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<IntListProto>, I>>(base?: I): IntListProto {
    return IntListProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IntListProto>, I>>(object: I): IntListProto {
    const message = createBaseIntListProto();
    message.values = object.values?.map((e) => e) || [];
    return message;
  },
};
// ...
export const IntMatrixProto: MessageFns<IntMatrixProto> = {
  encode(message: IntMatrixProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.values) {
      IntListProto.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): IntMatrixProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseIntMatrixProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.values.push(IntListProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<IntMatrixProto>, I>>(base?: I): IntMatrixProto {
    return IntMatrixProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<IntMatrixProto>, I>>(object: I): IntMatrixProto {
    const message = createBaseIntMatrixProto();
    message.values = object.values?.map((e) => IntListProto.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const EntityProto: MessageFns<EntityProto> = {
  encode(message: EntityProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.urn !== 0) {
      writer.uint32(8).int32(message.urn);
    }
    if (message.entityID !== 0) {
      writer.uint32(16).int32(message.entityID);
    }
    for (const v of message.properties) {
      PropertyProto.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EntityProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEntityProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.urn = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.entityID = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.properties.push(PropertyProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<EntityProto>, I>>(base?: I): EntityProto {
    return EntityProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EntityProto>, I>>(object: I): EntityProto {
    const message = createBaseEntityProto();
    message.urn = object.urn ?? 0;
    message.entityID = object.entityID ?? 0;
    message.properties = object.properties?.map((e) => PropertyProto.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const ChangeSetProto: MessageFns<ChangeSetProto> = {
  encode(message: ChangeSetProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.changes) {
      ChangeSetProto_EntityChangeProto.encode(v!, writer.uint32(10).fork()).join();
    }
    writer.uint32(18).fork();
    for (const v of message.deletes) {
      writer.int32(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChangeSetProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChangeSetProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.changes.push(ChangeSetProto_EntityChangeProto.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag === 16) {
            message.deletes.push(reader.int32());

            continue;
          }

          if (tag === 18) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.deletes.push(reader.int32());
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ChangeSetProto>, I>>(base?: I): ChangeSetProto {
    return ChangeSetProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChangeSetProto>, I>>(object: I): ChangeSetProto {
    const message = createBaseChangeSetProto();
    message.changes = object.changes?.map((e) => ChangeSetProto_EntityChangeProto.fromPartial(e)) || [];
    message.deletes = object.deletes?.map((e) => e) || [];
    return message;
  },
};

function createBaseChangeSetProto_EntityChangeProto(): ChangeSetProto_EntityChangeProto {
  return { entityID: 0, urn: 0, properties: [] };
}

export const ChangeSetProto_EntityChangeProto: MessageFns<ChangeSetProto_EntityChangeProto> = {
  encode(message: ChangeSetProto_EntityChangeProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.entityID !== 0) {
      writer.uint32(8).int32(message.entityID);
    }
    if (message.urn !== 0) {
      writer.uint32(16).int32(message.urn);
    }
    for (const v of message.properties) {
      PropertyProto.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChangeSetProto_EntityChangeProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChangeSetProto_EntityChangeProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.entityID = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.urn = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.properties.push(PropertyProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ChangeSetProto_EntityChangeProto>, I>>(
    base?: I,
  ): ChangeSetProto_EntityChangeProto {
    return ChangeSetProto_EntityChangeProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChangeSetProto_EntityChangeProto>, I>>(
    object: I,
  ): ChangeSetProto_EntityChangeProto {
    const message = createBaseChangeSetProto_EntityChangeProto();
    message.entityID = object.entityID ?? 0;
    message.urn = object.urn ?? 0;
    message.properties = object.properties?.map((e) => PropertyProto.fromPartial(e)) || [];
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}